+++ src/shared/ui/Components/CustomImage/ui/CustomImage.tsx
@@ -1,0 +1,14 @@
+import { Image, StyleProp, ImageStyle, ImageSourcePropType, ImageProps } from "react-native";
+import React from "react";
+import { styles } from "./CustomImage.styles";
+
+interface ICustomImage extends ImageProps {
+    path: ImageSourcePropType;
+    style?: StyleProp<ImageStyle>
+};
+
+export const CustomImage: React.FC<ICustomImage> = ({ path, style, ...props }) => {
+	return (
+		<Image {...props} style={[styles.checkedImg, style]} source={path} />
+	);
+};
